requirements.txt
langchain
langchain_core
langchain_community
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
langchain_openai
argparse
docx2txt
streamlit
langchainhub
langchaingroq
sentence-transformers
langchain_chroma
langchain_experimental